Elaborated on some comments, preparing to add collision detection
I elaborated on some of the comments so they are more educational if we
look at them later on. I also put the drawing algorithm inside a draw
loop, and set it to an interval. I also took out the radius from the
ball drawing arguments, and replaced it with the variable ballradius.
More variables will have to be added in order to handle collisions,
including position x and y, and velocity x and y.